requirements.txt
pbr>=0.5.21,<1.0
cliff>=1.4.3
keyring>=1.6.1,<2.0
pycrypto>=2.6
python-glanceclient>=0.9.0
python-keystoneclient>=0.4.1
python-novaclient>=2.15.0
python-cinderclient>=1.0.6
python-neutronclient>=2.3.0,<3
requests>=1.1
six>=1.4.1
